+++ utils-test-app/testsite/celery.py
@@ -15,3 +15,13 @@ app.config_from_object("django.conf:settings")
 
 # Load task modules from all registered Django app configs.
 app.autodiscover_tasks(lambda: settings.INSTALLED_APPS)
+
+# setup priorities ( 0 Highest, 9 Lowest )
+app.conf.broker_transport_options = {
+    "priority_steps": list(range(10)),  # setup que to have 10 steps
+    "queue_order_strategy": "priority",  # setup que to use prio sorting
+}
+
+# Automatically try to establish the connection to the AMQP broker on
+# Celery startup if it is unavailable.
+app.conf.broker_connection_retry_on_startup = True