+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [1.26.0] - 2025-04-14
+
+### Added
+
+- Class for Bootstrap styles: `views.BootstrapStyleBS5`
+- Adds support for Python 3.12
+
+### Changed
+
+- Deprecated: `views.BootstrapStyle`
+- Deprecated: `views.fontawesome_modal_button_html`
+
 ## [1.25.0] - 2023-12-14
 
 ### Added